font: sync underline_strikeout_metrics() with more recent version of font
Sigh, I really need to break out font.c/font.h to a separate
library...

+++ font.h
@@ -47,13 +47,13 @@ struct font {
     int max_x_advance;
 
     struct {
-        int position;
-        int thickness;
+        double position;
+        double thickness;
     } underline;
 
     struct {
-        int position;
-        int thickness;
+        double position;
+        double thickness;
     } strikeout;
 
     hash_entry_t **cache;
